Misc updates (#25)
* DiagonalPreconditioner: Make sure diagonal is dense

Since diag(::SparseMatrixCSC) return a SparseVector it is better to
convert to a dense vector with faster indexing. The diagonal should
always be dense for matrices where a diagonal preconditioner is used
anyway, so this does not affect the memory usage.

* DiagonalPreconditioner: fixes for offset arrays

This makes sure at-inbounds is only used with valid indices.